requirements.txt
fastapi
uvicorn
python-multipart
jinja2
watchdog
faster-whisper
pyyaml
pydantic
python-dotenv
